Fixing memory leak when removing tracks / streams
The _senderMap keeps references to no more used MediaStream / MediaTracks.
In an application where many tracks will be added/removed from a remote peer,
the "dead" tracks keep accumulating in _senderMap.

To avoid this, we turn _senderMap and its inner map into WeakMap.
This way, if the track is never used anymore, it will be freed from memory.